restructure klima taler and integrate it into to the run class
